Move ldap auth to LDAP::User. Removed unused code
